requirements.txt
fastapi
uvicorn
urllib3
wheel
motor
pydantic
requests
python-dotenv